[Update] #11 New experiment
There are two changes in this commit.
The first change involves the removal of the progress bar in DictLayer.py and ISTALayer.py to improve time performance.

The second change is in ISTALayer, where the 'h' parameter is now initialized with ones instead of zeroes, which improves loss during model training